requirements.txt
asgiref==3.7.2
dj-rest-auth==5.0.2
Django==5.0.1
django-debug-toolbar==4.2.0
django-filter==23.5
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==69.0.3
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
